% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretize_jointly.R
\name{cluster}
\alias{cluster}
\title{Cluster Multivariate Data}
\usage{
cluster(data, k, method)
}
\arguments{
\item{data}{input continuous multivariate data}

\item{k}{the number(s) of clusters}

\item{method}{the method for clustering}
}
\description{
The function obtains clusters from data using the given
  number of clusters, which may be a range.
}
