% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{gen_simdata}
\alias{gen_simdata}
\title{Generate Simulated Data}
\usage{
gen_simdata(cord, sim_table, noise = 0.3, plot = FALSE)
}
\arguments{
\item{cord}{data matrix that records the index for
each cluster on each dimension}

\item{sim_table}{a matrix}

\item{noise}{a numeric value to specify noise level}

\item{plot}{a logical to turn on or off plotting}
}
\description{
Generate Simulated Data
}
\keyword{internal}
