% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patterns.R
\name{plot.GridOnClusters}
\alias{plot.GridOnClusters}
\title{Plotting the continuous data along with cluster-preserving Grid}
\usage{
\method{plot}{GridOnClusters}(
  x,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  main.table = NULL,
  sub = NULL,
  pch = 19,
  ...
)
}
\arguments{
\item{x}{the result generated by discretize.jointly}

\item{xlab}{the horizontal axis label}

\item{ylab}{the vertical axis label}

\item{main}{the title of the clustering scatter plots}

\item{main.table}{the title of the discretized data plots}

\item{sub}{the subtitle}

\item{pch}{the symbol for points on the scatter plots}

\item{...}{additional graphical parameters}
}
\description{
Plots examples of jointly discretizing continuous data based on grids that 
preserve clusters in the original data.
}
