% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patterns.R
\name{plotGOCpatterns}
\alias{plotGOCpatterns}
\title{Plotting the continuous data along with cluster-preserving Grid}
\usage{
plotGOCpatterns(data, res)
}
\arguments{
\item{data}{a matrix containing two or more continuous variables.
Columns are variables, rows are observations.}

\item{res}{the result generated by discretize.jointly}
}
\description{
Plots examples of jointly discretizing continuous data based on grids that 
preserve clusters in the original data.
}
