% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster}
\alias{cluster}
\title{Cluster Multivariate Data}
\usage{
cluster(data, k, method, noise)
}
\arguments{
\item{data}{input continuous multivariate data}

\item{k}{the number(s) of clusters}

\item{method}{the method for clustering}

\item{noise}{adding jitter noise to the data or not}
}
\description{
The function obtains clusters from data using the given
  number of clusters, which may be a range.
}
\keyword{internal}
