% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patterns.R
\name{plot.GridOnClusters}
\alias{plot.GridOnClusters}
\title{Plotting Grid on Continuous Data}
\usage{
\method{plot}{GridOnClusters}(
  x,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  main.table = NULL,
  col,
  line_col = "black",
  cex = 1.125,
  sub = NULL,
  pch = 19,
  plot.table = TRUE,
  ...
)
}
\arguments{
\item{x}{the result generated by discretize.jointly}

\item{xlab}{the horizontal axis label}

\item{ylab}{the vertical axis label}

\item{main}{the title of the clustering scatter plots}

\item{main.table}{the title of the discretized data plots}

\item{col}{the color of data points}

\item{line_col}{the color of grid lines}

\item{cex}{A numerical value giving the amount by 
which plotting text and symbols should be magnified 
relative to the default.}

\item{sub}{the subtitle}

\item{pch}{the symbol for points on the scatter plots}

\item{plot.table}{a logical to show the contingency 
table. Default: \code{TRUE}.}

\item{...}{additional graphical parameters}
}
\description{
Plots discretized data based on grid that 
preserves clusters in original data.
}
