% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_patterns.R
\name{plotGOCpatterns}
\alias{plotGOCpatterns}
\title{Deprecated: Please use \code{plot()} instead}
\usage{
plotGOCpatterns(data, res)
}
\arguments{
\item{data}{the input continuous data matrix}

\item{res}{the result generated by discretize.jointly}
}
\description{
Plots examples of jointly discretizing continuous data based on grids that 
preserve clusters in the original data.
}
\keyword{internal}
